+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed container users to avoid permissions issues when using additional inventory
+
+### Changed
+- Changed image customization to allow additional inventory, while adding an automatic limit to Ansible
+
 ## [1.15.0] - 7/19/22
 ### Changed
 - Updated build references to AEE 1.3.0 (gitversion/giflow conversion).